Added an easter egg for a shotgun to be in the toilet

+++ Assets/Scripts/Upgrade Scripts/WeaponUpgrade.cs
@@ -7,7 +7,7 @@ public class WeaponUpgrade : MonoBehaviour
     public int firstUpgradePrice;
     public int secondUpgradePrice;
     public int thirdUpgradePrice;
-    public int distance;
+    public float distance;
 
     public int firstUpgradeDamage;
     public int thirdUpgradeDamage;
